refactor: move Table to Labs
