+++ packages/odyssey-react-labs/src/StaticTable.tsx
