+++ packages/odyssey-react-labs/src/index.ts
@@ -13,6 +13,10 @@
 export { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFns";
 export { LocalizationProvider } from "@mui/x-date-pickers";
 export type { LocalizationProviderProps } from "@mui/x-date-pickers";
+export type { MRT_ColumnDef as TableColumn } from "material-react-table";
 
 export * from "./DatePicker";
 export * from "./datePickerTheme";
+export * from "./materialReactTableTypes";
+export * from "./PaginatedTable";
+export * from "./StaticTable";
