+++ packages/odyssey-react-mui/package.json
@@ -30,7 +30,6 @@
     "@mui/utils": "^5.11.2",
     "@okta/odyssey-design-tokens": "workspace:*",
     "i18next": "^22.4.15",
-    "material-react-table": "^1.14.0",
     "react-i18next": "^12.2.2"
   },
   "devDependencies": {
