+++ packages/odyssey-storybook/src/components/odyssey-labs/DataTable/StaticTable.stories.tsx
@@ -11,19 +11,18 @@
  */
 
 import { Meta, StoryObj } from "@storybook/react";
+import { Button } from "@okta/odyssey-react-mui";
 import {
-  Button,
+  StaticTable,
+  StaticTableProps,
   TableColumn,
-  PaginatedTable,
-  PaginatedTableProps,
-} from "@okta/odyssey-react-mui";
+} from "@okta/odyssey-react-labs";
 
 import { MuiThemeDecorator } from "../../../../.storybook/components";
-import { useCallback, useRef, useState } from "react";
 
 const storybookMeta: Meta = {
-  title: "MUI Components/Table/Paginated",
-  component: PaginatedTable,
+  title: "Labs Components/Table/Static",
+  component: StaticTable,
   argTypes: {
     columns: {
       control: "array",
@@ -60,9 +59,6 @@ const storybookMeta: Meta = {
     },
     state: {
       control: "object",
-    },
-    ToolbarButtons: {
-      control: "function",
     },
   },
   decorators: [MuiThemeDecorator],
@@ -508,7 +504,7 @@ const data: Person[] = [
   },
 ];
 
-export const BasicUsage: StoryObj<PaginatedTableProps<Person>> = {
+export const BasicUsage: StoryObj<StaticTableProps<Person>> = {
   args: {
     columns,
     data,
@@ -516,38 +512,7 @@ export const BasicUsage: StoryObj<PaginatedTableProps<Person>> = {
   },
 };
 
-export const Pagination: StoryObj<PaginatedTableProps<Person>> = {
-  args: {
-    columns,
-    data,
-    getRowId: ({ id }: { id: string }) => id,
-  },
-  render: function C(args) {
-    const countRef = useRef(15);
-    const [data, setData] = useState(args.data.slice(0, countRef.current));
-
-    const fetchMoreData = useCallback(() => {
-      countRef.current = countRef.current + 10;
-
-      setData(args.data.slice(0, Math.min(countRef.current, args.data.length)));
-    }, [args.data]);
-
-    return (
-      <PaginatedTable {...args} data={data} fetchMoreData={fetchMoreData} />
-    );
-  },
-};
-
-export const Selection: StoryObj<PaginatedTableProps<Person>> = {
-  args: {
-    columns,
-    data,
-    getRowId: ({ id }: { id: string }) => id,
-    hasRowSelection: true,
-  },
-};
-
-export const CustomToolbar: StoryObj<PaginatedTableProps<Person>> = {
+export const CustomToolbar: StoryObj<StaticTableProps<Person>> = {
   args: {
     columns,
     data,
@@ -561,143 +526,3 @@ export const CustomToolbar: StoryObj<PaginatedTableProps<Person>> = {
     ),
   },
 };
-
-// const reportColumns: TableColumn<any>[] = [
-//   {
-//     accessorKey: "name",
-//     Cell: ({ cell }) => <Link href="#none">{cell.getValue<string>()}</Link>,
-//     header: "Name",
-//   },
-//   {
-//     accessorFn: (object) => object.authenticators,
-//     Cell: ({ cell }) => cell.getValue<string[]>().join(", "),
-//     header: "Authenticators",
-//   },
-//   {
-//     accessorKey: "createdDate",
-//     Cell: ({ cell }) => new Date(cell.getValue<string>()).getFullYear(),
-//     header: "Address",
-//   },
-// ];
-
-// const reportData: any[] = [
-//   {
-//     login: "login_VCXLG_2@okta.com",
-//     email: "testemail_VCXLG_2@okta.com",
-//     name: "TestFirstName_2 TestLastName_2",
-//     primaryPhone: null,
-//     mobilePhone: "555-415-1337",
-//     phone: "555-415-1337",
-//     status: "ACTIVE",
-//     isAdmin: false,
-//     authenticators: ["Email", "Password"],
-//     authenticatorsCount: 2,
-//     groupNames: ["ReportsTestGroup_VCXLG"],
-//     createdDate: "2021-06-27T06:11:27.000Z",
-//     activationDate: "2021-06-27T06:11:29.000Z",
-//   },
-//   {
-//     login: "login_VCXLG_4@okta.com",
-//     email: "testemail_VCXLG_4@okta.com",
-//     name: "TestFirstName_4 TestLastName_4",
-//     primaryPhone: null,
-//     mobilePhone: "555-415-1337",
-//     phone: "555-415-1337",
-//     status: "ACTIVE",
-//     isAdmin: false,
-//     authenticators: ["Email", "Password"],
-//     authenticatorsCount: 2,
-//     groupNames: ["ReportsTestGroup_VCXLG"],
-//     createdDate: "2021-06-27T06:11:40.000Z",
-//     activationDate: "2021-06-27T06:11:40.000Z",
-//   },
-//   {
-//     login: "login_VCXLG_5@okta.com",
-//     email: "testemail_VCXLG_5@okta.com",
-//     name: "TestFirstName_5 TestLastName_5",
-//     primaryPhone: null,
-//     mobilePhone: "555-415-1337",
-//     phone: "555-415-1337",
-//     status: "ACTIVE",
-//     isAdmin: false,
-//     authenticators: ["Email", "Password"],
-//     authenticatorsCount: 2,
-//     groupNames: ["ReportsTestGroup_VCXLG"],
-//     createdDate: "2021-06-27T06:11:44.000Z",
-//     activationDate: "2021-06-27T06:11:44.000Z",
-//   },
-//   {
-//     login: "login_VCXLG_6@okta.com",
-//     email: "testemail_VCXLG_6@okta.com",
-//     name: "TestFirstName_6 TestLastName_6",
-//     primaryPhone: null,
-//     mobilePhone: "555-415-1337",
-//     phone: "555-415-1337",
-//     status: "ACTIVE",
-//     isAdmin: false,
-//     authenticators: ["Email", "Password"],
-//     authenticatorsCount: 2,
-//     groupNames: ["ReportsTestGroup_VCXLG"],
-//     createdDate: "2021-06-27T06:11:48.000Z",
-//     activationDate: "2021-06-27T06:11:48.000Z",
-//   },
-//   {
-//     login: "login_uEGLa_1@okta.com",
-//     email: "testemail_uEGLa_1@okta.com",
-//     name: "Johnny Unverified",
-//     primaryPhone: null,
-//     mobilePhone: "555-415-1337",
-//     phone: "555-415-1337",
-//     status: "ACTIVE",
-//     isAdmin: false,
-//     authenticators: ["Email", "Password"],
-//     authenticatorsCount: 2,
-//     groupNames: ["ReportsTestGroup_uEGLa"],
-//     createdDate: "2021-06-27T06:26:57.000Z",
-//     activationDate: "2021-06-27T06:26:57.000Z",
-//   },
-// ];
-
-/*
-- Column visibility view w/ filter box at the top because there are so many columns.
-- Remove context menu on column headers.
-- Add title to header.
-- Remove header.
-- Add `enableRowVirtualization`.
-- Add `onColumnVisibilityChange`.
-- Hide firstName column by default `initialState={{ columnVisibility: { firstName: false } }}` PREFERRED TO HAVE HIDDEN PROP ON COLUMN.
-- Pass component to expanded row. -- NOT TODAY
-- Infinite scroll.
-*/
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-// export const Reports: StoryObj<TableProps<any>> = {
-//   args: {
-//     columns: reportColumns,
-//     data: reportData,
-//   },
-// };
-
-/*
-- Pagination, but no known item count.
-- Sorting
-- Custom Sorting
-- Row Selection
-- Custom Cell Render
-- More buttons (3) and search on left
-- Search makes an API call
-- Other filters outside the table
-- Know which rows are selected
-- Customized LOADING state -> Progress bar on top of table for loading table data. Want to keep showing old items while loading.
-- Customized ERROR state -> "Something went wrong" and "Reload" button to refetch data.
-- Customized NO_DATA state -> "No pending review items found" when searching and nothing's there.
-- Change size (CSS `flex`, `text-align`, `header-align`) of table cells
-- Initial sort state
-- Sorting order (multi-sort)
-- Row re-ordering. Currently drag 'n drop, but we could also do "type in number" to change row order.
-- Expandable row icon on right side
-- Custom cell and header types
-  - Text Alignment
-  - `MenuButton`
-  - Date display
-- `StaticTable` with virtualization, no pagination, and no footer display nor toolbars.
-*/
