+++ packages/odyssey-storybook/src/components/odyssey-mui/Icon/Icon.stories.tsx
@@ -17,9 +17,8 @@ import {
   type IconProps,
   iconDictionary,
   iconSizeValues,
-  StaticTable,
-  type TableColumn,
 } from "@okta/odyssey-react-mui";
+import { StaticTable, type TableColumn } from "@okta/odyssey-react-labs";
 import { MuiThemeDecorator } from "../../../../.storybook/components/MuiThemeDecorator";
 
 const storybookMeta: Meta<IconProps> = {
